requirements.txt
pandas>=2.0.0
openai>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
python-dotenv>=1.0.0
selenium>=4.0.0
webdriver-manager>=3.8.0 